+++ R/cor_test_freq.R
@@ -5,8 +5,14 @@
   var_x <- .complete_variable_x(data, x, y)
   var_y <- .complete_variable_y(data, x, y)
 
+  .cor_test_base(x, y, var_x, var_y, ci = ci, method = method, ...)
+}
+
+
+#' @keywords internal
+.cor_test_base <- function(x, y, var_x, var_y, ci = 0.95, method = "pearson", ...) {
   method <- match.arg(tolower(method), c("pearson", "kendall", "spearman"), several.ok = FALSE)
-  rez <- stats::cor.test(var_x, var_y, conf.level = ci, method = method, alternative = "two.sided", exact = FALSE)
+  rez <- stats::cor.test(var_x, var_y, conf.level = ci, method = method, exact = FALSE, ...)
 
   params <- parameters::model_parameters(rez)
   params$Parameter1 <- x
@@ -19,8 +25,8 @@
   }
 
   # Add CI for non-pearson correlations
-  if (method %in% c( "kendall", "spearman")){
-    rez_ci <- cor_to_ci(rez$estimate, n=length(var_x), ci=ci, ...)
+  if (method %in% c("kendall", "spearman")) {
+    rez_ci <- cor_to_ci(rez$estimate, n = length(var_x), ci = ci, ...)
     params$CI_low <- rez_ci$CI_low
     params$CI_high <- rez_ci$CI_high
   }